+++ CHANGELOG.md
@@ -73,6 +73,7 @@ Also, there are quite a few offloading improvements and many quality-of-life cha
   - add `--new` command line flag to enable testing of new packages without breaking existing installs  
   - downgrade rocm to `torch==2.7.1`  
   - set the minimum supported rocm version on linux to `rocm==6.0`  
+  - disallow `zluda` and `directml` on non-windows platforms  
 - **OpenVINO**
   - update to `openvino==2025.3.0`  
   - add deprecation warning for Python 3.9